chore: Update comment and stock controllers to use route attributes

+++ api/Controllers/CommentController.cs
@@ -30,6 +30,10 @@ namespace api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+
+            if (ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var comments = await _commentRepository.GetAllAsync();
             var commentDto = comments.Select(c => c.MapToCommentDto());
             return Ok(comments);
@@ -39,6 +43,10 @@ namespace api.Controllers
         [Route("{id:int}")]
         public async Task<IActionResult> GetCommentById([FromRoute] int id)
         {
+
+            if (ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var comment = await _commentRepository.GetByIdAsync(id);
             if (comment == null)
             {
@@ -53,6 +61,9 @@ namespace api.Controllers
         [Route("{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, [FromBody] CreateCommentDto createCommentDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             if (!await _stockRepository.StockExists(stockId))
             {
                 return BadRequest(new { message = "Stock does not exist" });
@@ -60,7 +71,7 @@ namespace api.Controllers
 
             var commentModel = createCommentDto.MapToCommentFromCreateDto(stockId);
             await _commentRepository.CreateAsync(commentModel);
-            
+
             var createdComment = commentModel.MapToCommentDto();
             return CreatedAtAction(nameof(GetCommentById), new { id = createdComment.Id }, createdComment);
         }
@@ -68,7 +79,12 @@ namespace api.Controllers
 
         [HttpPut]
         [Route("{id:int}")]
-        public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentDto updateCommentDto){
+        public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentDto updateCommentDto)
+        {
+
+            if (ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var commentModel = await _commentRepository.UpdateAsync(id, updateCommentDto);
 
             if (commentModel == null)
@@ -83,6 +99,10 @@ namespace api.Controllers
         [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+
+            if (ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var commentModel = await _commentRepository.DeleteAsync(id);
 
             if (commentModel == null)
